backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0
httpx==0.27.0
openai==1.55.0
supabase==2.8.0
pytest==8.3.0
pytest-asyncio==0.24.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10 